test.h作成とMakefile修正
