+++ sandbox/hash.c
