+++ sandbox/pointer.c
