+++ test/parser/test_is_target_file_extension.c
@@ -5,14 +5,6 @@
 // ここにis_target_file_extension関数を含めるか、別のヘッダファイルからインクルードします。
 bool	is_target_file_extension(const char *filename, const char *extension);
 
-void setUp(void) {
-    // ここにはテストのセットアップに必要なコードを記述します。
-}
-
-void tearDown(void) {
-    // ここにはテストのクリーンアップに必要なコードを記述します。
-}
-
 void test_is_target_file_extension_with_correct_extension(void) {
     TEST_ASSERT_TRUE(is_target_file_extension("test.txt", ".txt"));
 }
@@ -41,8 +33,8 @@ void test_is_target_file_extension_with_included_but_incorrect_extension(void) {
     TEST_ASSERT_FALSE(is_target_file_extension("test.txt.jpg", ".txt"));
 }
 
-int main(void) {
-    UNITY_BEGIN();
+void test_is_target_file_extension(void)
+{
     RUN_TEST(test_is_target_file_extension_with_correct_extension);
     RUN_TEST(test_is_target_file_extension_without_extension);
     RUN_TEST(test_is_target_file_extension_with_null_filename);
@@ -50,5 +42,5 @@ int main(void) {
     RUN_TEST(test_is_target_file_extension_with_incorrect_start_of_extension);
     RUN_TEST(test_is_target_file_extension_with_filename_starting_dot);
     RUN_TEST(test_is_target_file_extension_with_included_but_incorrect_extension);
-    return UNITY_END();
+	
 }
