+++ test/test.h
@@ -1,0 +1,8 @@
+#ifndef TEST_H
+# define TEST_H
+
+#include "unity.h"
+
+void test_is_target_file_extension(void);
+
+#endif
